Struct gdbstub::arch::lldb::Register

pub struct Register<'a> {
    pub name: &'a str,
    pub alt_name: Option<&'a str>,
    pub bitsize: usize,
    pub offset: usize,
    pub encoding: Encoding,
    pub format: Format,
    pub set: &'a str,
    pub gcc: Option<usize>,
    pub dwarf: Option<usize>,
    pub generic: Option<Generic>,
    pub container_regs: Option<&'a [usize]>,
    pub invalidate_regs: Option<&'a [usize]>,
}

Describes the register info for a single register of the target.

Fields
name: &'a str

The primary register name.

alt_name: Option<&'a str>

An alternate name for the register.

bitsize: usize

Size in bits of a register.

offset: usize

The offset within the ‘g’ and ‘G’ packet of the register data for this register.

encoding: Encoding

The encoding type of the register.

format: Format

The preferred format for display of this register.

set: &'a str

The register set name this register belongs to.

gcc: Option<usize>

The GCC compiler registers number for this register.

Note: This denotes the same KEY:VALUE; pair as ehframe:VALUE;. See the LLDB source.

dwarf: Option<usize>

The DWARF register number for this register that is used for this register in the debug information.

generic: Option<Generic>

Specify as a generic register.

container_regs: Option<&'a [usize]>

Other concrete register values this register is contained in.

invalidate_regs: Option<&'a [usize]>

Specifies which register values should be invalidated when this register is modified.