Module gdbstub::stub

The core GdbStub type, used to drive a GDB debugging session for a particular Target over a given Connection.

Modules

run_blocking

Types and traits related to the GdbStub::run_blocking interface.

state_machine

Low-level state-machine interface that underpins GdbStub.

Structs

GdbStub

Debug a Target using the GDB Remote Serial Protocol over a given Connection.

GdbStubBuilder

Helper to construct and customize GdbStub.

Enums

BaseStopReason

Describes why a thread stopped.

DisconnectReason

Describes why the GDB session ended.

GdbStubBuilderError

An error which may occur when building a GdbStub.

GdbStubError

An error which may occur during a GDB debugging session.

Traits

IntoStopReason

A marker trait implemented by SingleThreadStopReason and MultiThreadStopReason.

Type Definitions

MultiThreadStopReason

A stop reason for a multi threaded target.

SingleThreadStopReason

A stop reason for a single threaded target.