Module gdbstub::stub::state_machine

Low-level state-machine interface that underpins GdbStub.

Hey, what gives? Where are all the docs!?

Yep, sorry about that!

gdbstub 0.6 turned out ot be a pretty massive release, and documenting everything has proven to be a somewhat gargantuan task that’s kept delaying the release data further and further back…

To avoid blocking the release any further, I’ve decided to leave this bit of the API sparsely documented.

If you’re interested in using this API directly (e.g: to integrate gdbstub into a no_std project, or to use gdbstub in a non-blocking manner alongside async/await / a project specific event loop), your best bet would be to review the following bits of code to get a feel for the API:

The implementation of GdbStub::run_blocking

Implementations of BlockingEventLoop used alongside GdbStub::run_blocking (e.g: the in-tree armv4t / armv4t_multicore examples)

Real-world projects using the API

The best example of this (at the time of writing) is the code at vmware-labs/node-replicated-kernel

If you have any questions, feel free to open a discussion thread over at the gdbstub GitHub repo.

Modules

state

State machine typestates.

Structs

GdbStubStateMachineInner

Core state machine implementation that is parameterized by various states. Can be converted back into the appropriate GdbStubStateMachine variant via Into::into.

Enums

GdbStubStateMachine

State-machine interface to GdbStub.