Struct gdbstub::stub::state_machine::GdbStubStateMachineInner

pub struct GdbStubStateMachineInner<'a, S, T: Target, C: Connection> { /* private fields */ }

Core state machine implementation that is parameterized by various states. Can be converted back into the appropriate GdbStubStateMachine variant via Into::into.

Implementations

impl<'a, S, T: Target, C: Connection> GdbStubStateMachineInner<'a, S, T, C>

Methods which can be called regardless of the current state.

pub fn borrow_conn(&mut self) -> &mut C

Return a mutable reference to the underlying connection.

impl<'a, T: Target, C: Connection> GdbStubStateMachineInner<'a, Idle<T>, T, C>

Methods which can only be called from the GdbStubStateMachine::Idle state.

pub fn incoming_data(
self,
target: &mut T,
byte: u8
) -> Result<GdbStubStateMachine<'a, T, C>, Error<T::Error, C::Error>>

Pass a byte to the GDB stub.

impl<'a, T: Target, C: Connection> GdbStubStateMachineInner<'a, Running, T, C>

Methods which can only be called from the GdbStubStateMachine::Running state.

pub fn report_stop(
self,
target: &mut T,
reason: impl IntoStopReason<T>
) -> Result<GdbStubStateMachine<'a, T, C>, Error<T::Error, C::Error>>

Report a target stop reason back to GDB.

pub fn incoming_data(
self,
target: &mut T,
byte: u8
) -> Result<GdbStubStateMachine<'a, T, C>, Error<T::Error, C::Error>>

Pass a byte to the GDB stub.

NOTE: unlike the incoming_data method in the state::Idle state, this method does not perform any state transitions, and will return a GdbStubStateMachineInner in the state::Running state.

impl<'a, T: Target, C: Connection> GdbStubStateMachineInner<'a, CtrlCInterrupt, T, C>

Methods which can only be called from the GdbStubStateMachine::CtrlCInterrupt state.

pub fn interrupt_handled(
self,
target: &mut T,
stop_reason: Option<impl IntoStopReason<T>>
) -> Result<GdbStubStateMachine<'a, T, C>, Error<T::Error, C::Error>>

Acknowledge the Ctrl-C interrupt.

Passing None as a stop reason will return the state machine to whatever state it was in pre-interruption, without immediately returning a stop reason.

Depending on how the target is implemented, it may or may not make sense to immediately return a stop reason as part of handling the Ctrl-C interrupt. e.g: in some cases, it may be better to send the target a signal upon receiving a Ctrl-C interrupt without immediately sending a stop reason, and instead deferring the stop reason to some later point in the target’s execution.

Some notes on handling Ctrl-C interrupts:

Stubs are not required to recognize these interrupt mechanisms, and the precise meaning associated with receipt of the interrupt is implementation defined.

If the target supports debugging of multiple threads and/or processes, it should attempt to interrupt all currently-executing threads and processes.

If the stub is successful at interrupting the running program, it should send one of the stop reply packets (see Stop Reply Packets) to GDB as a result of successfully stopping the program

impl<'a, T: Target, C: Connection> GdbStubStateMachineInner<'a, Disconnected, T, C>

Methods which can only be called from the GdbStubStateMachine::Disconnected state.

pub fn get_reason(&self) -> DisconnectReason

Inspect why the GDB client disconnected.

pub fn return_to_idle(self) -> GdbStubStateMachine<'a, T, C>

Reuse the existing state machine instance, reentering the idle loop.