Module gdbstub::target::ext::base

Base operations required to debug any target (read/write memory/registers, step/resume, etc…)

It is recommended that single threaded targets implement the simplified singlethread API, as gdbstub includes optimized implementations of certain internal routines when operating in singlethreaded mode.

Modules

multithread

Base debugging operations for multi threaded targets.

singlethread

Base debugging operations for single threaded targets.

Structs

GdbInterrupt

A handle to check for incoming GDB interrupts.

GdbInterruptNoAsync

A simplified interface to GdbInterrupt for projects without async/await infrastructure.

Enums

BaseOps

Base operations for single/multi threaded targets.

ReplayLogPosition

Describes the point reached in a replay log for the corresponding stop reason.

ResumeAction

Describes how the target should be resumed.

Traits

SingleRegisterAccess

Target Extension - Support for single-register access.

Type Definitions

SingleRegisterAccessOps

See SingleRegisterAccess