Trait gdbstub::target::ext::base::multithread::MultiThreadOps

pub trait MultiThreadOps: Target {

    fn resume(
        &mut self, 
        default_resume_action: ResumeAction, 
        gdb_interrupt: GdbInterrupt<'_>
    ) -> Result<ThreadStopReason<<Self::Arch as Arch>::Usize>, Self::Error>;

    fn clear_resume_actions(&mut self) -> Result<(), Self::Error>;

    fn set_resume_action(
        &mut self, 
        tid: Tid, 
        action: ResumeAction
    ) -> Result<(), Self::Error>;

    fn read_registers(
        &mut self, 
        regs: &mut <Self::Arch as Arch>::Registers, 
        tid: Tid
    ) -> TargetResult<(), Self>;

    fn write_registers(
        &mut self, 
        regs: &<Self::Arch as Arch>::Registers, 
        tid: Tid
    ) -> TargetResult<(), Self>;

    fn read_addrs(
        &mut self, 
        start_addr: <Self::Arch as Arch>::Usize, 
        data: &mut [u8], 
        tid: Tid
    ) -> TargetResult<(), Self>;

    fn write_addrs(
        &mut self, 
        start_addr: <Self::Arch as Arch>::Usize, 
        data: &[u8], 
        tid: Tid
    ) -> TargetResult<(), Self>;

    fn list_active_threads(
        &mut self, 
        thread_is_active: &mut dyn FnMut(Tid)
    ) -> Result<(), Self::Error>;

    fn support_range_step(
        &mut self
    ) -> Option<MultiThreadRangeSteppingOps<'_, Self>> { ... }

    fn support_reverse_step(
        &mut self
    ) -> Option<MultiThreadReverseStepOps<'_, Self>> { ... }

    fn support_reverse_cont(
        &mut self
    ) -> Option<MultiThreadReverseContOps<'_, Self>> { ... }

    fn single_register_access(
        &mut self
    ) -> Option<SingleRegisterAccessOps<'_, Tid, Self>> { ... }

    fn is_thread_alive(&mut self, tid: Tid) -> Result<bool, Self::Error> { ... }

}

Base debugging operations for multi threaded targets.

Required methods

fn resume(
&mut self,
default_resume_action: ResumeAction,
gdb_interrupt: GdbInterrupt<'_>
) -> Result<ThreadStopReason<<Self::Arch as Arch>::Usize>, Self::Error>

Resume execution on the target.

Prior to calling resume, gdbstub will call clear_resume_actions, followed by zero or more calls to set_resume_action, specifying any thread-specific resume actions.

The default_action parameter specifies the “fallback” resume action for any threads that did not have a specific resume action set via set_resume_action. The GDB client typically sets this to ResumeAction::Continue, though this is not guaranteed.

The check_gdb_interrupt callback can be invoked to check if GDB sent an Interrupt packet (i.e: the user pressed Ctrl-C). It’s recommended to invoke this callback every-so-often while the system is running (e.g: every X cycles/milliseconds). Periodically checking for incoming interrupt packets is not required, but it is recommended.

Implementation requirements

These requirements cannot be satisfied by gdbstub internally, and must be handled on a per-target basis.

Adjusting PC after a breakpoint is hit

The GDB remote serial protocol documentation notes the following:

On some architectures, such as x86, at the architecture level, when a breakpoint instruction executes the program counter points at the breakpoint address plus an offset. On such targets, the stub is responsible for adjusting the PC to point back at the breakpoint address.

Omitting PC adjustment may result in unexpected execution flow and/or breakpoints not working correctly.

Additional Considerations

Bare-Metal Targets

On bare-metal targets (such as microcontrollers or emulators), it’s common to treat individual CPU cores as a separate “threads”. e.g: in a dual-core system, [CPU0, CPU1] might be mapped to [TID1, TID2] (note that TIDs cannot be zero).

In this case, the Tid argument of read/write_addrs becomes quite relevant, as different cores may have different memory maps.

Running in “Non-stop” mode

At the moment, gdbstub only supports GDB’s “All-Stop” mode, whereby all threads must be stopped when returning from resume (not just the thread associated with the ThreadStopReason).

fn clear_resume_actions(&mut self) -> Result<(), Self::Error>

Clear all previously set resume actions.

fn set_resume_action(
&mut self,
tid: Tid,
action: ResumeAction
) -> Result<(), Self::Error>

Specify what action each thread should take when resume is called.

A simple implementation of this method would simply update an internal HashMap<Tid, ResumeAction>.

Aside from the four “base” resume actions handled by this method (i.e: Step, Continue, StepWithSignal, and ContinueWithSignal), there are also two additional resume actions which are only set if the target implements their corresponding protocol extension:

Action	Protocol Extension

Optimized Range Stepping	See support_range_step()

“Stop”	Used in “Non-Stop” mode *

* “Non-Stop” mode is currently unimplemented

fn read_registers(
&mut self,
regs: &mut <Self::Arch as Arch>::Registers,
tid: Tid
) -> TargetResult<(), Self>

Read the target’s registers.

If the registers could not be accessed, an appropriate non-fatal error should be returned.

fn write_registers(
&mut self,
regs: &<Self::Arch as Arch>::Registers,
tid: Tid
) -> TargetResult<(), Self>

Write the target’s registers.

If the registers could not be accessed, an appropriate non-fatal error should be returned.

fn read_addrs(
&mut self,
start_addr: <Self::Arch as Arch>::Usize,
data: &mut [u8],
tid: Tid
) -> TargetResult<(), Self>

Read bytes from the specified address range.

If the requested address range could not be accessed (e.g: due to MMU protection, unhanded page fault, etc…), an appropriate non-fatal error should be returned.

fn write_addrs(
&mut self,
start_addr: <Self::Arch as Arch>::Usize,
data: &[u8],
tid: Tid
) -> TargetResult<(), Self>

Write bytes to the specified address range.

If the requested address range could not be accessed (e.g: due to MMU protection, unhanded page fault, etc…), an appropriate non-fatal error should be returned.

fn list_active_threads(
&mut self,
thread_is_active: &mut dyn FnMut(Tid)
) -> Result<(), Self::Error>

List all currently active threads.

See the section above on implementing thread-related methods on bare-metal (threadless) targets.

Provided methods

fn support_range_step(
&mut self
) -> Option<MultiThreadRangeSteppingOps<'_, Self>>

Support for the optimized range stepping resume action.

fn support_reverse_step(
&mut self
) -> Option<MultiThreadReverseStepOps<'_, Self>>

Support for reverse stepping a target.

fn support_reverse_cont(
&mut self
) -> Option<MultiThreadReverseContOps<'_, Self>>

Support for reverse continuing a target.

fn single_register_access(
&mut self
) -> Option<SingleRegisterAccessOps<'_, Tid, Self>>

Support for single-register access. See SingleRegisterAccess for more details.

While this is an optional feature, it is highly recommended to implement it when possible, as it can significantly improve performance on certain architectures.

fn is_thread_alive(&mut self, tid: Tid) -> Result<bool, Self::Error>

Check if the specified thread is alive.

As a convenience, this method provides a default implementation which uses list_active_threads to do a linear-search through all active threads. On thread-heavy systems, it may be more efficient to override this method with a more direct query.