Trait gdbstub::target::ext::base::multithread::MultiThreadRangeStepping

pub trait MultiThreadRangeStepping: Target + MultiThreadResume {
    fn set_resume_action_range_step(
        &mut self, 
        tid: Tid, 
        start: <Self::Arch as Arch>::Usize, 
        end: <Self::Arch as Arch>::Usize
    ) -> Result<(), Self::Error>;
}

Target Extension - Optimized range stepping for multi threaded targets. See MultiThreadResume::support_range_step.

Required methods

fn set_resume_action_range_step(
&mut self,
tid: Tid,
start: <Self::Arch as Arch>::Usize,
end: <Self::Arch as Arch>::Usize
) -> Result<(), Self::Error>

Range step the specified target thread.

Range Stepping will step the target once, and keep stepping the target as long as execution remains between the specified start (inclusive) and end (exclusive) addresses, or another stop condition is met (e.g: a breakpoint it hit).

If the range is empty (start == end), then the action becomes equivalent to the ‘s’ action. In other words, single-step once, and report the stop (even if the stepped instruction jumps to start).

Note: A stop reply may be sent at any point even if the PC is still within the stepping range; for example, it is valid to implement range stepping in a degenerate way as a single instruction step operation.