Trait gdbstub::target::ext::base::multithread::MultiThreadResume

pub trait MultiThreadResume: Target {
    fn resume(&mut self) -> Result<(), Self::Error>;

    fn clear_resume_actions(&mut self) -> Result<(), Self::Error>;

    fn set_resume_action_continue(
        &mut self, 
        tid: Tid, 
        signal: Option<Signal>
    ) -> Result<(), Self::Error>;

    fn support_single_step(
        &mut self
    ) -> Option<MultiThreadSingleStepOps<'_, Self>> { ... }

    fn support_range_step(
        &mut self
    ) -> Option<MultiThreadRangeSteppingOps<'_, Self>> { ... }

    fn support_reverse_step(&mut self) -> Option<ReverseStepOps<'_, Tid, Self>> { ... }

    fn support_reverse_cont(&mut self) -> Option<ReverseContOps<'_, Tid, Self>> { ... }
}

Target extension - support for resuming multi threaded targets.

Required methods

fn resume(&mut self) -> Result<(), Self::Error>

Resume execution on the target.

Prior to calling resume, gdbstub will call clear_resume_actions, followed by zero or more calls to the set_resume_action_XXX methods, specifying any thread-specific resume actions.

Upon returning from the resume method, the target being debugged should be configured to run according to whatever resume actions the GDB client had specified using any of the set_resume_action_XXX methods.

Any thread that wasn’t explicitly resumed by a set_resume_action_XXX method should be resumed as though it was resumed with set_resume_action_continue.

A basic target implementation only needs to implement support for set_resume_action_continue, with all other resume actions requiring their corresponding protocol extension to be implemented:

Action	Protocol Extension

Optimized Single Stepping	See support_single_step()

Optimized Range Stepping	See support_range_step()

“Stop”	Used in “Non-Stop” mode *

* “Non-Stop” mode is currently unimplemented in gdbstub

Additional Considerations

Adjusting PC after a breakpoint is hit

The GDB remote serial protocol documentation notes the following:

On some architectures, such as x86, at the architecture level, when a breakpoint instruction executes the program counter points at the breakpoint address plus an offset. On such targets, the stub is responsible for adjusting the PC to point back at the breakpoint address.

Omitting PC adjustment may result in unexpected execution flow and/or breakpoints not appearing to work correctly.

Bare-Metal Targets

On bare-metal targets (such as microcontrollers or emulators), it’s common to treat individual CPU cores as a separate “threads”. e.g: in a dual-core system, [CPU0, CPU1] might be mapped to [TID1, TID2] (note that TIDs cannot be zero).

In this case, the Tid argument of read/write_addrs becomes quite relevant, as different cores may have different memory maps.

fn clear_resume_actions(&mut self) -> Result<(), Self::Error>

Clear all previously set resume actions.

fn set_resume_action_continue(
&mut self,
tid: Tid,
signal: Option<Signal>
) -> Result<(), Self::Error>

Continue the specified thread.

See the resume docs for information on when this is called.

The GDB client may also include a signal which should be passed to the target.

Provided methods

fn support_single_step(&mut self) -> Option<MultiThreadSingleStepOps<'_, Self>>

Support for optimized single stepping.

fn support_range_step(
&mut self
) -> Option<MultiThreadRangeSteppingOps<'_, Self>>

Support for optimized range stepping.

fn support_reverse_step(&mut self) -> Option<ReverseStepOps<'_, Tid, Self>>

Support for reverse stepping a target.

fn support_reverse_cont(&mut self) -> Option<ReverseContOps<'_, Tid, Self>>

Support for reverse continuing a target.