Trait gdbstub::target::ext::base::multithread::MultiThreadReverseCont

pub trait MultiThreadReverseCont: Target + MultiThreadOps {
    fn reverse_cont(
        &mut self, 
        gdb_interrupt: GdbInterrupt<'_>
    ) -> Result<ThreadStopReason<<Self::Arch as Arch>::Usize>, Self::Error>;
}

Target Extension - Reverse continue for multi threaded targets.

Reverse continue allows the target to run backwards until it reaches the end of the replay log.

Required methods

fn reverse_cont(
&mut self,
gdb_interrupt: GdbInterrupt<'_>
) -> Result<ThreadStopReason<<Self::Arch as Arch>::Usize>, Self::Error>

Reverse-continue the target.