Enum gdbstub::target::ext::base::singlethread::StopReason

#[non_exhaustive]pub enum StopReason<U> {
    DoneStep,
    GdbInterrupt,
    Exited(u8),
    Terminated(u8),
    Signal(u8),
    SwBreak,
    HwBreak,
    Watch {
        kind: WatchKind,
        addr: U,
    },
    ReplayLog(ReplayLogPosition),
}

Describes why the target stopped.

Targets MUST only respond with stop reasons that correspond to IDETs that target has implemented.

e.g: A target which has not implemented the HwBreakpoint IDET must not return a HwBreak stop reason. While this is not enforced at compile time, doing so will result in a runtime UnsupportedStopReason error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

DoneStep

Completed the single-step request.

GdbInterrupt

check_gdb_interrupt returned true.

Exited(u8)

The process exited with the specified exit status.

Terminated(u8)

The process terminated with the specified signal number.

Signal(u8)

The program received a signal.

SwBreak

Hit a software breakpoint (e.g. due to a trap instruction).

Requires: SwBreakpoint.

NOTE: This does not necessarily have to be a breakpoint configured by the client/user of the current GDB session.

HwBreak

Hit a hardware breakpoint.

Requires: HwBreakpoint.

Watch

Hit a watchpoint.

Requires: HwWatchpoint.

Fields of Watch

kind: WatchKind

Kind of watchpoint that was hit

addr: U

Address of watched memory

ReplayLog(ReplayLogPosition)

The program has reached the end of the logged replay events.

Requires: SingleThreadReverseCont or SingleThreadReverseStep.

This is used for GDB's reverse execution. When playing back a recording, you may hit the end of the buffer of recorded events, and as such no further execution can be done. This stop reason tells GDB that this has occurred.