Trait gdbstub::target::ext::base::singlethread::SingleThreadOps

pub trait SingleThreadOps: Target {
    fn resume(
        &mut self, 
        action: ResumeAction, 
        gdb_interrupt: GdbInterrupt<'_>
    ) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>;

    fn read_registers(
        &mut self, 
        regs: &mut <Self::Arch as Arch>::Registers
    ) -> TargetResult<(), Self>;

    fn write_registers(
        &mut self, 
        regs: &<Self::Arch as Arch>::Registers
    ) -> TargetResult<(), Self>;

    fn read_addrs(
        &mut self, 
        start_addr: <Self::Arch as Arch>::Usize, 
        data: &mut [u8]
    ) -> TargetResult<(), Self>;

    fn write_addrs(
        &mut self, 
        start_addr: <Self::Arch as Arch>::Usize, 
        data: &[u8]
    ) -> TargetResult<(), Self>;

    fn support_resume_range_step(
        &mut self
    ) -> Option<SingleThreadRangeSteppingOps<'_, Self>> { ... }

    fn support_reverse_step(
        &mut self
    ) -> Option<SingleThreadReverseStepOps<'_, Self>> { ... }

    fn support_reverse_cont(
        &mut self
    ) -> Option<SingleThreadReverseContOps<'_, Self>> { ... }

    fn single_register_access(
        &mut self
    ) -> Option<SingleRegisterAccessOps<'_, (), Self>> { ... }
}

Base debugging operations for single threaded targets.

Required methods

fn resume(
&mut self,
action: ResumeAction,
gdb_interrupt: GdbInterrupt<'_>
) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>

Resume execution on the target.

action specifies how the target should be resumed (i.e: step or continue).

The check_gdb_interrupt callback can be invoked to check if GDB sent an Interrupt packet (i.e: the user pressed Ctrl-C). It’s recommended to invoke this callback every-so-often while the system is running (e.g: every X cycles/milliseconds). Periodically checking for incoming interrupt packets is not required, but it is recommended.

Implementation requirements

These requirements cannot be satisfied by gdbstub internally, and must be handled on a per-target basis.

Adjusting PC after a breakpoint is hit

The GDB remote serial protocol documentation notes the following:

On some architectures, such as x86, at the architecture level, when a breakpoint instruction executes the program counter points at the breakpoint address plus an offset. On such targets, the stub is responsible for adjusting the PC to point back at the breakpoint address.

Omitting PC adjustment may result in unexpected execution flow and/or breakpoints not appearing to work correctly.

fn read_registers(
&mut self,
regs: &mut <Self::Arch as Arch>::Registers
) -> TargetResult<(), Self>

Read the target’s registers.

fn write_registers(
&mut self,
regs: &<Self::Arch as Arch>::Registers
) -> TargetResult<(), Self>

Write the target’s registers.

fn read_addrs(
&mut self,
start_addr: <Self::Arch as Arch>::Usize,
data: &mut [u8]
) -> TargetResult<(), Self>

Read bytes from the specified address range.

If the requested address range could not be accessed (e.g: due to MMU protection, unhanded page fault, etc…), an appropriate non-fatal error should be returned.

fn write_addrs(
&mut self,
start_addr: <Self::Arch as Arch>::Usize,
data: &[u8]
) -> TargetResult<(), Self>

Write bytes to the specified address range.

If the requested address range could not be accessed (e.g: due to MMU protection, unhanded page fault, etc…), an appropriate non-fatal error should be returned.

Provided methods

fn support_resume_range_step(
&mut self
) -> Option<SingleThreadRangeSteppingOps<'_, Self>>

Support for the optimized range stepping resume action.

fn support_reverse_step(
&mut self
) -> Option<SingleThreadReverseStepOps<'_, Self>>

Support for reverse stepping a target.

fn support_reverse_cont(
&mut self
) -> Option<SingleThreadReverseContOps<'_, Self>>

Support for reverse continuing a target.

fn single_register_access(
&mut self
) -> Option<SingleRegisterAccessOps<'_, (), Self>>

Support for single-register access. See SingleRegisterAccess for more details.

While this is an optional feature, it is highly recommended to implement it when possible, as it can significantly improve performance on certain architectures.