Trait gdbstub::target::ext::base::singlethread::SingleThreadRangeStepping

pub trait SingleThreadRangeStepping: Target + SingleThreadOps {
    fn resume_range_step(
        &mut self, 
        start: <Self::Arch as Arch>::Usize, 
        end: <Self::Arch as Arch>::Usize, 
        gdb_interrupt: GdbInterrupt<'_>
    ) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>;
}

Target Extension - Optimized range stepping for single threaded targets. See SingleThreadOps::support_resume_range_step.

Range Stepping will step the target once, and keep stepping the target as long as execution remains between the specified start (inclusive) and end (exclusive) addresses, or another stop condition is met (e.g: a breakpoint it hit).

If the range is empty (start == end), then the action becomes equivalent to the ‘s’ action. In other words, single-step once, and report the stop (even if the stepped instruction jumps to start).

Note: A stop reply may be sent at any point even if the PC is still within the stepping range; for example, it is valid to implement range stepping in a degenerate way as a single instruction step operation.

Required methods

fn resume_range_step(
&mut self,
start: <Self::Arch as Arch>::Usize,
end: <Self::Arch as Arch>::Usize,
gdb_interrupt: GdbInterrupt<'_>
) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>

See SingleThreadOps::resume.