Trait gdbstub::target::ext::base::singlethread::SingleThreadResume

pub trait SingleThreadResume: Target {
    fn resume(&mut self, signal: Option<Signal>) -> Result<(), Self::Error>;

    fn support_single_step(
        &mut self
    ) -> Option<SingleThreadSingleStepOps<'_, Self>> { ... }

    fn support_range_step(
        &mut self
    ) -> Option<SingleThreadRangeSteppingOps<'_, Self>> { ... }

    fn support_reverse_step(&mut self) -> Option<ReverseStepOps<'_, (), Self>> { ... }

    fn support_reverse_cont(&mut self) -> Option<ReverseContOps<'_, (), Self>> { ... }
}

Target extension - support for resuming single threaded targets.

Required methods

fn resume(&mut self, signal: Option<Signal>) -> Result<(), Self::Error>

Resume execution on the target.

The GDB client may also include a signal which should be passed to the target.

Additional Considerations

Adjusting PC after a breakpoint is hit

The GDB remote serial protocol documentation notes the following:

On some architectures, such as x86, at the architecture level, when a breakpoint instruction executes the program counter points at the breakpoint address plus an offset. On such targets, the stub is responsible for adjusting the PC to point back at the breakpoint address.

Omitting PC adjustment may result in unexpected execution flow and/or breakpoints not appearing to work correctly.

Provided methods

fn support_single_step(&mut self) -> Option<SingleThreadSingleStepOps<'_, Self>>

Support for optimized single stepping.

fn support_range_step(
&mut self
) -> Option<SingleThreadRangeSteppingOps<'_, Self>>

Support for optimized range stepping.

fn support_reverse_step(&mut self) -> Option<ReverseStepOps<'_, (), Self>>

Support for reverse stepping a target.

fn support_reverse_cont(&mut self) -> Option<ReverseContOps<'_, (), Self>>

Support for reverse continuing a target.