Trait gdbstub::target::ext::base::singlethread::SingleThreadReverseCont

pub trait SingleThreadReverseCont: Target + SingleThreadOps {
    fn reverse_cont(
        &mut self, 
        gdb_interrupt: GdbInterrupt<'_>
    ) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>;
}

Target Extension - Reverse continue for single threaded targets.

Reverse continue allows the target to run backwards until it reaches the end of the replay log.

Required methods

fn reverse_cont(
&mut self,
gdb_interrupt: GdbInterrupt<'_>
) -> Result<StopReason<<Self::Arch as Arch>::Usize>, Self::Error>

Reverse-continue the target.

Loading content...