Trait gdbstub::target::ext::breakpoints::HwBreakpoint

pub trait HwBreakpoint: Target + Breakpoints {
    fn add_hw_breakpoint(
        &mut self, 
        addr: <Self::Arch as Arch>::Usize, 
        kind: <Self::Arch as Arch>::BreakpointKind
    ) -> TargetResult<bool, Self>;

    fn remove_hw_breakpoint(
        &mut self, 
        addr: <Self::Arch as Arch>::Usize, 
        kind: <Self::Arch as Arch>::BreakpointKind
    ) -> TargetResult<bool, Self>;
}

Nested Target Extension - Set/Remove Hardware Breakpoints.

See this stackoverflow discussion about the differences between hardware and software breakpoints.

Recommendation: If you’re implementing Target for an emulator that’s using an interpreted CPU (as opposed to a JIT), there shouldn’t be any reason to implement this extension (as software breakpoints are likely to be just-as-fast).

Required methods

fn add_hw_breakpoint(
&mut self,
addr: <Self::Arch as Arch>::Usize,
kind: <Self::Arch as Arch>::BreakpointKind
) -> TargetResult<bool, Self>

Add a new hardware breakpoint.

Return Ok(false) if the operation could not be completed.

fn remove_hw_breakpoint(
&mut self,
addr: <Self::Arch as Arch>::Usize,
kind: <Self::Arch as Arch>::BreakpointKind
) -> TargetResult<bool, Self>

Remove an existing hardware breakpoint.

Return Ok(false) if the operation could not be completed.