Module gdbstub::target::ext::extended_mode

Enables Extended Mode functionality when connecting using target extended-remote, such as spawning new processes and/or attaching to existing processes.

Disclaimer

While this API has been end-to-end tested and confirmed working with a “toy” target implementation (see the included armv4t example), it has not been “battle-tested” with a fully-featured extended-mode capable target.

If you end up using this API to implement an extended-mode capable target, please file an issue on the repo detailing any bugs / usability issues you may encountered while implementing this API! If everything happens to Just Work as expected, nonetheless file an issue so that this disclaimer can be removed in future releases!

Structs

Args

Iterator of args passed to a spawned process (used in ExtendedMode::run)

Enums

AttachKind

Describes how the target attached to a process.

ShouldTerminate

Returned from ExtendedMode::kill

Traits

ConfigureAslr

Nested Target Extension - Enable/Disable ASLR for spawned processes (for a more consistent debugging experience).

ConfigureEnv

Nested Target Extension - Set/Remove/Reset the Environment variables for spawned processes.

ConfigureStartupShell

Nested Target Extension - Configure if spawned processes should be spawned using a shell.

ConfigureWorkingDir

Nested Target Extension - Configure the working directory for spawned processes.

ExtendedMode

Target Extension - Support Extended Mode functionality.

Type Definitions

ConfigureAslrOps

See ConfigureAslr.

ConfigureEnvOps

See ConfigureEnv.

ConfigureStartupShellOps

See ConfigureStartupShell.

ConfigureWorkingDirOps

See ConfigureWorkingDir.

ExtendedModeOps

See ExtendedMode.