Enum gdbstub::target::ext::host_io::HostIoErrno

pub enum HostIoErrno {

    EPERM,
    ENOENT,
    EINTR,
    EBADF,
    EACCES,
    EFAULT,
    EBUSY,
    EEXIST,
    ENODEV,
    ENOTDIR,
    EISDIR,
    EINVAL,
    ENFILE,
    EMFILE,
    EFBIG,
    ENOSPC,
    ESPIPE,
    EROFS,
    ENAMETOOLONG,
    EUNKNOWN,

}

Errno values for Host I/O operations.

Extracted from the GDB documentation at https://sourceware.org/gdb/onlinedocs/gdb/Errno-Values.html

Variants

EPERM

Operation not permitted (POSIX.1-2001).

ENOENT

No such file or directory (POSIX.1-2001).

Typically, this error results when a specified pathname does not exist, or one of the components in the directory prefix of a pathname does not exist, or the specified pathname is a dangling symbolic link.

EINTR

Interrupted function call (POSIX.1-2001); see signal(7).

EBADF

Bad file descriptor (POSIX.1-2001).

EACCES

Permission denied (POSIX.1-2001).

EFAULT

Bad address (POSIX.1-2001).

EBUSY

Device or resource busy (POSIX.1-2001).

EEXIST

File exists (POSIX.1-2001).

ENODEV

No such device (POSIX.1-2001).

ENOTDIR

Not a directory (POSIX.1-2001).

EISDIR

Is a directory (POSIX.1-2001).

EINVAL

Invalid argument (POSIX.1-2001).

ENFILE

Too many open files in system (POSIX.1-2001). On Linux, this is probably a result of encountering the /proc/sys/fs/file-max limit (see proc(5)).

EMFILE

Too many open files (POSIX.1-2001). Commonly caused by exceeding the RLIMIT_NOFILE resource limit described in getrlimit(2).

EFBIG

File too large (POSIX.1-2001).

ENOSPC

No space left on device (POSIX.1-2001).

ESPIPE

Invalid seek (POSIX.1-2001).

EROFS

Read-only filesystem (POSIX.1-2001).

ENAMETOOLONG

Filename too long (POSIX.1-2001).

EUNKNOWN

Unknown errno - there may not be a GDB mapping for this value