Trait gdbstub::target::ext::host_io::HostIoPread

pub trait HostIoPread: HostIo {
    fn pread(
        &mut self, 
        fd: u32, 
        count: usize, 
        offset: u64, 
        buf: &mut [u8]
    ) -> HostIoResult<usize, Self>;
}

Nested Target Extension - Host I/O pread operation.

Required methods

fn pread(
&mut self,
fd: u32,
count: usize,
offset: u64,
buf: &mut [u8]
) -> HostIoResult<usize, Self>

Read data from the open file corresponding to fd.

Up to count bytes will be read from the file, starting at offset relative to the start of the file.

Return the number of bytes written into buf (which may be less than count).

If offset is greater than the length of the underlying data, return Ok(0).