Module gdbstub::target::ext::section_offsets

Get section/segment relocation offsets from the target.

For some targets, sections may be relocated from their base address. As a result, the stub may need to tell GDB the final section addresses to ensure that debug symbols are resolved correctly after relocation.

Note: This extension corresponds to the qOffsets command, which is limited to reporting the offsets for code, data and bss, and is generally considered a legacy feature.

For targets where library offsets are maintained externally (e.g. Windows) you should consider implementing the more flexible qXfer:library:read. See issue #20 for more info.

For System-V architectures GDB is capable of extracting library offsets from memory if it knows the base address of the dynamic linker. The base address can be specified by either implementing this command or by including a AT_BASE entry in the response to the more modern qXfer:auxv:read command. See issue #20 for more info.

Enums

Offsets

Describes the offset the target loaded the image sections at, so the target can notify GDB that it needs to adjust the addresses of symbols.

Traits

SectionOffsets

Target Extension - Get section/segment relocation offsets from the target.

Type Definitions

SectionOffsetsOps

See SectionOffsets.