Module gdbstub::arch

Traits to encode architecture-specific target information.

Community created Arch Implementations

Before getting your hands dirty and implementing a new Arch from scratch, make sure to check out gdbstub_arch, a companion crate to gdbstub which aggregates community-created Arch implementations for most common architectures!

Note: Prior to gdbstub 0.5, Arch implementations were distributed as a part of the main gdbstub crate (under the gdbstub::arch module). This wasn't ideal, any gdbstub::arch-level breaking-changes forced the entire gdbstub crate to release a new (potentially breaking!) version.

Having community-created Arch implementations distributed in a separate crate helps minimize any unnecessary "version churn" in gdbstub core.

Traits

Arch
Encodes architecture-specific information, such as pointer size, register layout, etc...

BreakpointKind
Breakpoint kind for specific architectures.

RegId
Register identifier for target registers.

Registers
Methods to read/write architecture-specific registers.