Trait gdbstub::arch::BreakpointKind

pub trait BreakpointKind: Sized + Debug {
    fn from_usize(kind: usize) -> Option<Self>;
}

Breakpoint kind for specific architectures.

This trait corresponds to the kind field of the “z” and “Z” breakpoint packets, as documented here.

A breakpoint “kind” is architecture-specific and typically indicates the size of the breakpoint in bytes that should be inserted. As such, most architectures will set BreakpointKind = usize.

Some architectures, such as ARM and MIPS, have additional meanings for kind. See the Architecture-Specific Protocol Details section of the GBD documentation for more details.

If no architecture-specific value is being used, kind should be set to ‘0’, and the BreakpointKind associated type should be ().

Required methods

fn from_usize(kind: usize) -> Option<Self>

Parse Self from a raw usize.