Trait gdbstub::target::ext::breakpoints::HwWatchpoint

pub trait HwWatchpoint: Target + Breakpoints {
    fn add_hw_watchpoint(
        &mut self, 
        addr: <Self::Arch as Arch>::Usize, 
        kind: WatchKind
    ) -> TargetResult<bool, Self>;

    fn remove_hw_watchpoint(
        &mut self, 
        addr: <Self::Arch as Arch>::Usize, 
        kind: WatchKind
    ) -> TargetResult<bool, Self>;
}

Nested Target Extension - Set/Remove Hardware Watchpoints.

See the GDB documentation regarding watchpoints for how they’re supposed to work.

Note: If this extension isn’t implemented, GDB will default to using software watchpoints, which tend to be excruciatingly slow (as hey are implemented by single-stepping the system, and reading the watched memory location after each step).

Required methods

fn add_hw_watchpoint(
&mut self,
addr: <Self::Arch as Arch>::Usize,
kind: WatchKind
) -> TargetResult<bool, Self>

Add a new hardware watchpoint. Return Ok(false) if the operation could not be completed.

fn remove_hw_watchpoint(
&mut self,
addr: <Self::Arch as Arch>::Usize,
kind: WatchKind
) -> TargetResult<bool, Self>

Remove an existing hardware watchpoint. Return Ok(false) if the operation could not be completed.