Trait gdbstub::target::ext::extended_mode::ConfigureEnv

pub trait ConfigureEnv: ExtendedMode {
    fn set_env(
        &mut self, 
        key: &[u8], 
        val: Option<&[u8]>
    ) -> TargetResult<(), Self>;

    fn remove_env(&mut self, key: &[u8]) -> TargetResult<(), Self>;

    fn reset_env(&mut self) -> TargetResult<(), Self>;
}

Nested Target Extension - Set/Remove/Reset the Environment variables for spawned processes.

Corresponds to GDB's set environment cmd.

Note: Environment variables are not guaranteed to be UTF-8, and are passed as raw byte arrays. If the provided keys/values could not be converted into an appropriate representation, a non fatal error should be returned.

Required methods

fn set_env(&mut self, key: &[u8], val: Option<&[u8]>) -> TargetResult<(), Self>

Set an environment variable.

fn remove_env(&mut self, key: &[u8]) -> TargetResult<(), Self>

Remove an environment variable.

fn reset_env(&mut self) -> TargetResult<(), Self>

Reset all environment variables to their initial state (i.e: undo all previous set/remove_env calls).

Loading content...