Trait gdbstub::target::ext::extended_mode::ExtendedMode

pub trait ExtendedMode: Target {
    fn run(
        &mut self, 
        filename: Option<&[u8]>, 
        args: Args<'_, '_>
    ) -> TargetResult<Pid, Self>;

    fn attach(&mut self, pid: Pid) -> TargetResult<(), Self>;

    fn query_if_attached(&mut self, pid: Pid) -> TargetResult<AttachKind, Self>;

    fn kill(&mut self, pid: Option<Pid>) -> TargetResult<ShouldTerminate, Self>;

    fn restart(&mut self) -> Result<(), Self::Error>;

    fn on_start(&mut self) -> Result<(), Self::Error> { ... }

    fn configure_aslr(&mut self) -> Option<ConfigureAslrOps<'_, Self>> { ... }

    fn configure_env(&mut self) -> Option<ConfigureEnvOps<'_, Self>> { ... }

    fn configure_startup_shell(
        &mut self
    ) -> Option<ConfigureStartupShellOps<'_, Self>> { ... }

    fn configure_working_dir(
        &mut self
    ) -> Option<ConfigureWorkingDirOps<'_, Self>> { ... }
}

Target Extension - Support Extended Mode functionality.

Extended Mode for Single/Multi Threaded Targets

While extended-mode is primarily intended to be implemented by targets which support debugging multiple processes, there's no reason why a basic single/multi-threaded target can't implement these extensions as well.

For example, instead of "spawning" a process, the run command could be used to reset the execution state instead (e.g: resetting an emulator).

Required methods

fn run(
&mut self,
filename: Option<&[u8]>,
args: Args<'_, '_>
) -> TargetResult<Pid, Self>

Spawn and attach to the program filename, passing it the provided args on its command line.

The program is created in the stopped state.

If no filename is provided, the stub may use a default program (e.g. the last program run), or a non fatal error should be returned.

filename and args are not guaranteed to be valid UTF-8, and are passed as raw byte arrays. If the filenames/arguments could not be converted into an appropriate representation, a non fatal error should be returned.

Note: This method's implementation should handle any additional configuration options set via the various ConfigureXXX extensions to ExtendedMode. e.g: if the ConfigureEnv extension is implemented and enabled, this method should set the spawned processes' environment variables accordingly.

fn attach(&mut self, pid: Pid) -> TargetResult<(), Self>

Attach to a new process with the specified PID.

In all-stop mode, all threads in the attached process are stopped; in non-stop mode, it may be attached without being stopped (if that is supported by the target).

fn query_if_attached(&mut self, pid: Pid) -> TargetResult<AttachKind, Self>

Query if specified PID was spawned by the target (via run), or if the target attached to an existing process (via attach).

If the PID doesn't correspond to a process the target has run or attached to, a non fatal error should be returned.

fn kill(&mut self, pid: Option<Pid>) -> TargetResult<ShouldTerminate, Self>

Called when the GDB client sends a Kill request.

If the PID doesn't correspond to a process the target has run or attached to, a non fatal error should be returned.

GDB may or may not specify a specific PID to kill. When no PID is specified, the target is free to decide what to do (e.g: kill the last-used pid, terminate the connection, etc...).

If ShouldTerminate::Yes is returned, GdbStub will immediately stop and return a DisconnectReason::Kill. Otherwise, the connection will remain open, and GdbStub will continue listening for run/attach requests.

fn restart(&mut self) -> Result<(), Self::Error>

Restart the program being debugged.

The GDB docs don't do a good job describing what a "restart" operation entails. For reference, the official gdbserver seems to kill all inferior processes, and then re-run whatever program was provided on the command line (if one was provided).

Author's Note: Based on my current (as of Sept 2020) understanding of the GDB client;s source code, it seems that the "R" packet is never sent so-long as the target implements the "vRun" packet (which corresponds to this trait's run method). As such, while gdbstub exposes this functionality, and "requires" an implementation, unless you're running a fairly old version of GDB, it should be fine to simply stub it out -- e.g: using the unimplemented!() macro / returning a fatal error.

Loading content...
Provided methods

fn on_start(&mut self) -> Result<(), Self::Error>

(optional) Invoked when GDB client switches to extended mode.

The default implementation is a no-op.

Target implementations can override this implementation if they need to perform any operations once extended mode is activated.

fn configure_aslr(&mut self) -> Option<ConfigureAslrOps<'_, Self>>

Enable/Disable ASLR for spawned processes.

fn configure_env(&mut self) -> Option<ConfigureEnvOps<'_, Self>>

Set/Remove/Reset Environment variables for spawned processes.

fn configure_startup_shell(
&mut self
) -> Option<ConfigureStartupShellOps<'_, Self>>

Configure if spawned processes should be spawned using a shell.

fn configure_working_dir(&mut self) -> Option<ConfigureWorkingDirOps<'_, Self>>

Configure the working directory for spawned processes.

Loading content...