Trait gdbstub::target::Target

pub trait Target {
    type Arch: Arch;
    type Error;
    fn base_ops(&mut self) -> BaseOps<'_, Self::Arch, Self::Error>;

    fn breakpoints(&mut self) -> Option<BreakpointsOps<'_, Self>> { ... }

    fn monitor_cmd(&mut self) -> Option<MonitorCmdOps<'_, Self>> { ... }

    fn extended_mode(&mut self) -> Option<ExtendedModeOps<'_, Self>> { ... }

    fn section_offsets(&mut self) -> Option<SectionOffsetsOps<'_, Self>> { ... }

    fn target_description_xml_override(
        &mut self
    ) -> Option<TargetDescriptionXmlOverrideOps<'_, Self>> { ... }
}

Describes the architecture and capabilities of a target which can be debugged by GdbStub.

The Target trait describes how to control and modify a system’s execution state during a GDB debugging session, and serves as the primary bridge between gdbstub’s generic protocol implementation and a target’s project/platform-specific code.

Target is the most important trait in gdbstub, and must be implemented by anyone who uses the library!

Please refer to the the documentation in the target module for more information on how to implement and work with Target and its various extension traits.

Associated Types

type Arch: Arch

The target’s architecture.

type Error

A target-specific fatal error.

Required methods

fn base_ops(&mut self) -> BaseOps<'_, Self::Arch, Self::Error>

Base operations such as reading/writing from memory/registers, stopping/resuming the target, etc….

For example, on a single-threaded target:

use gdbstub::target::Target;
use gdbstub::target::base::singlethread::SingleThreadOps;

impl SingleThreadOps for MyTarget {
    // ...
}

impl Target for MyTarget {
    fn base_ops(&mut self) -> base::BaseOps<Self::Arch, Self::Error> {
        base::BaseOps::SingleThread(self)
    }
}

Provided methods

fn breakpoints(&mut self) -> Option<BreakpointsOps<'_, Self>>

Set/Remove software breakpoints.

fn monitor_cmd(&mut self) -> Option<MonitorCmdOps<'_, Self>>

Handle custom GDB monitor commands.

fn extended_mode(&mut self) -> Option<ExtendedModeOps<'_, Self>>

Support for Extended Mode operations.

fn section_offsets(&mut self) -> Option<SectionOffsetsOps<'_, Self>>

Handle requests to get the target’s current section (or segment) offsets.

fn target_description_xml_override(
&mut self
) -> Option<TargetDescriptionXmlOverrideOps<'_, Self>>

Override the target description XML specified by Target::Arch.